Initial search pkg restructure with ligo.p_astro
