--- src/spiir-data/spiir/data/__init__.py
@@ -1,1 +1,0 @@
-from . import distribution
