--- src/spiir-data/spiir/data/distribution/__init__.py
@@ -1,3 +1,0 @@
-from . import config, constraint, distribution, transform
-from .distribution import (Distribution, JointDistribution, NumPyDistribution,
-                           PyCBCDistribution)
