--- src/spiir-data/spiir/data/distribution/constraint/__init__.py
@@ -1,2 +1,0 @@
-from .constraint import (Constraint, NumPyConstraint, SPIIRConstraint,
-                         load_constraints_from_config)
