--- src/spiir-data/spiir/data/distribution/transform/__init__.py
@@ -1,2 +1,0 @@
-from .transform import (NumPyTransform, SPIIRTransform, Transform,
-                        load_transforms_from_config)
