--- src/spiir-io/spiir/io/__init__.py
@@ -1,1 +1,0 @@
-from . import array, cli, ligolw, logging, mp
