--- src/spiir-io/spiir/io/igwn/__init__.py
@@ -1,2 +1,0 @@
-from . import consumers
-from .consumer import run_igwn_alert_consumer
