--- src/spiir-io/spiir/io/igwn/consumers/__init__.py
@@ -1,1 +1,0 @@
-from .composite_mass_contour import FGMCMassContourConsumer
