--- src/spiir-io/spiir/io/ligolw/__init__.py
@@ -1,15 +1,0 @@
-from . import postcoh
-from .array import (append_psd_series_to_ligolw, append_snr_series_to_ligolw,
-                    build_psd_series_from_xmldoc, build_snr_series_from_xmldoc,
-                    get_array_from_xmldoc, get_arrays_from_xmldoc,
-                    load_arrays_from_xml, load_arrays_from_xmls,
-                    load_psd_series_from_xml, load_snr_series_from_xml)
-from .coinc import load_coinc_xml, save_coinc_xml
-from .ligolw import LIGOLWContentHandler, load_ligolw_xmldoc, strip_ilwdchar
-from .param import (append_p_astro_to_ligolw, get_p_astro_from_xmldoc,
-                    get_params_from_xmldoc, load_params_from_xml)
-from .table import (append_table_to_ligolw, build_dataframe_from_table,
-                    get_table_from_xmldoc, get_tables_from_xmldoc,
-                    load_ligolw_tables, load_table_from_xml,
-                    load_table_from_xmls, load_tables_from_xml,
-                    load_tables_from_xmls)
