--- src/spiir-search/spiir/search/__init__.py
@@ -1,1 +1,0 @@
-from . import p_astro
