--- src/spiir-search/spiir/search/p_astro/models/__init__.py
@@ -1,1 +1,0 @@
-from .composite import FGMCChirpMassAreaModel
