+++ src/spiir/search/p_astro/__init__.py
@@ -1,0 +1,1 @@
+from . import models, mchirp_area
